+++ docs/14_0_0/gettingstarted/whatsnew.md
@@ -25,6 +25,9 @@ Look into [migration guide](https://primefaces.github.io/primefaces/14_0_0/#/../
     * JPALazyDataModel now supports case insensitive filters with `setCaseSensitive(false);`
     * JPALazyDataModel now supports wildcard filters with `setWildcardSupport(true);` so you can use `*`, `%`, `_` or `?` in filter
     * JPALazyDataModel now supports builder pattern for constructor.
+
+* Wizard
+    * Added `disableOnAjax` attribute to disable next and back navigation buttons during Ajax requests triggered by them.
   
 ### PrimeFaces Selenium